connection between models has been fixed now you can add value to a task

+++ src/main/java/levente/libardi/abris/repository/UserRepository.java
@@ -3,5 +3,5 @@ package levente.libardi.abris.repository;
 import levente.libardi.abris.model.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<AppUser, Integer> {
+public interface UserRepository extends JpaRepository<AppUser, Long> {
 }
